requirements.txt
slack_bolt
Flask
uWSGI>=2,<3
python-decouple
aiohttp